[REFACTOR] remove duplicate logics for creating ticket type

+++ src/main/java/id/ac/ui/cs/advprog/eventspherre/service/TicketServiceImpl.java
@@ -1,6 +1,7 @@
 package id.ac.ui.cs.advprog.eventspherre.service;
 
 import id.ac.ui.cs.advprog.eventspherre.model.Ticket;
+import id.ac.ui.cs.advprog.eventspherre.model.User;
 import id.ac.ui.cs.advprog.eventspherre.repository.TicketRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -31,7 +32,10 @@ public class TicketServiceImpl implements TicketService {
     }
 
     @Override
-    public void deleteTicket(UUID id) {
+    public void deleteTicket(UUID id, User user) {
+        if (user.getRole() != User.Role.ADMIN) {
+            throw new IllegalArgumentException("Only admins can delete tickets");
+        }
         ticketRepository.deleteById(id);
     }
